requirements.txt
asgiref==3.7.2
awsebcli==3.20.10
botocore==1.31.85
cement==2.8.2
certifi==2024.2.2
channels==4.0.0
charset-normalizer==3.3.2
colorama==0.4.3
cycler==0.11.0
distlib==0.3.8
Django==3.2.25
djangorestframework==3.14.0
drf-yasg==1.21.7
EB==0.1.5
filelock==3.12.2
fonttools==4.38.0
future==1.0.0
idna==3.6
importlib-metadata==6.7.0
inflection==0.5.1
iso8601==2.1.0
jmespath==1.0.1
kiwisolver==1.4.5
matplotlib==3.5.3
mysqlclient @ file:///C:/Users/ee/downloads/mysqlclient-1.4.6-cp37-cp37m-win_amd64.whl
numpy==1.21.6
opencv-contrib-python==4.9.0.80
opencv-python==4.9.0.80
packaging==24.0
pathspec==0.10.1
Pillow==9.5.0
platformdirs==4.0.0
pyparsing==3.1.2
pypiwin32==223
pyserial==3.5
python-dateutil==2.9.0.post0
pytz==2024.1
pywin32==306
PyYAML==6.0.1
requests==2.31.0
scipy==1.7.3
semantic-version==2.8.5
serial==0.0.97
six==1.16.0
sqlparse==0.4.4
termcolor==1.1.0
typing_extensions==4.7.1
uritemplate==4.1.1
urllib3==1.26.18
virtualenv==20.26.2
wcwidth==0.1.9
zipp==3.15.0
